Persisted game state and bug fixes
Added a feature to persist the game state on exit. Whenever the game exits we save the current character stats, items, and floor. If the player is in combat and exits using a long hold on the back button, we also save the current monster type and health. This way, when reentering the game, we can go back to the fight. The persisted data also has a version number which can be incremented if the data we need to save ever changes.

Fixed some residual bugs from the upgrade to 2.0. This includes a memory leak of bitmaps and a bug where bitmaps would not show when a new window was first loaded.

Removed the exit confirmation window. The existence of persisted data makes it unnecessary.