Phase 1 of extending MiniDungeon
Rewriting the tech used for progression in order to generalize it and allow for more in depth adventures.
Location: Holds information about a location (dungeon floor?) including background art, monsters available, and how long it takes to traverse the floor
Story: Collects Locations and Monsters.
Adventure changes: Dungeon floors are now time based instead of random. Items no longer drop but can only be purchased (I may change this back later).
Known issues: Persistent data is off for now; the shop is available while in the dungeon; the end of the dungeon is completely untested; the win condition is currently off.