Adding rudimentary profiling
Profiling is off by default, but can be turned on in MiniDungeon.h. ProfileLogStart and ProfileLogStop cannot be nested due to statics. They also seem to give wildly varying results for operations. My current theory is that the time to wake up the watch may grossly overshadow the time to do any other operations.