Added option not to launch app on events
By default, turning on the background thread will cause the app to launch if an event happens. This new feature allows the used to turn that off. This way, time passes and an event might trigger while in other apps, but the game will not interrupt. The event will simply happen the next time the user launches MiniDungeon.